afs: Also search ipv4 address list in afs_find_server
When searching for a matching peer, look in fs_addresses4 list rather
than the fs_addresses6 list if the peer address is v4.

When processing an incoming callback from an ipv4 address, this would
lead to a failure to set call->server, resulting in the callback being
ignored and the client seeing stale contents.

Fixes: 72904d7b9bfb ("rxrpc, afs: Allow afs to pin rxrpc_peer objects")
Signed-off-by: Marc Dionne <marc.dionne@auristor.com>